+++ projects/console-web-front/src/components/projects/guides/MobileGuide.tsx
@@ -10,7 +10,7 @@ import Link from 'next/link';
 
 import { GuideSupportLanguage, mobileGuideData, SAMPLE_GIT_URL } from '../../../resources/guide';
 import { flexRowBaseStyle } from '../../../styles/box';
-import CopyButtonContainer from './CopyButtonContainer';
+import CopyButtonContainer from './CodeWithCopyButton';
 import useRequest from '../../../hooks/useRequest';
 import { uploadSampleApplication } from '../../../api/project-application';
 import { sendErrorNotification, sendSuccessNotification } from '../../../utils/antd';
